25.04.09.(수) 11:50 물품 등록 진행중(사진 업로드가 안되는 문제 해결 중)

+++ src/main/java/com/solomarket/entity/FileEntity.java
@@ -30,10 +30,10 @@ public class FileEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long fileNo;
+    private int fileNo;
 
     @Column(nullable = false)
-    private Long productNo;
+    private int productNo;
 
     @Column(nullable = false)
     private String fileName;
